requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiostream==0.5.2
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
atlassian-python-api==3.41.3
attrs==23.1.0
audioread==3.0.1
av==10.0.0
azure-cognitiveservices-speech==1.32.1
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
cryptography==41.0.3
ctranslate2==3.21.0
dataclasses-json==0.5.14
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
Eel==0.16.0
exceptiongroup==1.1.3
executing==2.0.1
faster-whisper==0.9.0
fastjsonschema==2.18.1
filelock==3.13.1
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
geojson==2.5.0
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==2.0.1
h11==0.14.0
html2text==2020.1.16
httpcore==0.16.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.335
langsmith==0.0.63
lazy_loader==0.3
librosa==0.10.1
llama-hub==0.0.44
llama-index==0.8.69.post2
llvmlite==0.41.1
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.2
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.7
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.2
openai==1.2.4
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pinecone-client==2.2.4
pip==23.3.1
platformdirs==4.0.0
pluggy==1.3.0
pooch==1.8.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.14.3
pydeck==0.8.1b0
Pygments==2.16.1
pyOpenSSL==23.2.0
pyowm==3.3.0
pyparsing==3.1.1
pypdf==3.17.0
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
pytest==7.4.3
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-telegram-bot==20.6
pytz==2023.3.post1
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
slave==0.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.21
stack-data==0.6.3
stringcase==1.2.0
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
transformers==4.35.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
wheel==0.41.2
whichcraft==0.6.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yarl==1.7.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
